+++ application/admin/controller/Base.php
@@ -9,12 +9,14 @@ namespace app\admin\controller;
 
 use com\verify\HonrayVerify;
 use app\common\controller\Common;
-use think\Request;
+use think\facade\Request;
 
 class Base extends Common
 {
     public function login()
     {   
+        // dump("OK");
+        // die();
         $userModel = model('User');
         $param = $this->param;
         $username = $param['username'];
@@ -85,11 +87,7 @@ class Base extends Common
     // miss 路由：处理没有匹配到的路由规则
     public function miss()
     {
-        if (Request::instance()->isOptions()) {
-            return ;
-        } else {
-            echo 'vuethink接口';
-        }
+        echo 'thinkvue OK';
     }
 }