+++ config/session.php
@@ -22,5 +22,5 @@ return [
     // 驱动方式 支持redis memcache memcached
     'type'           => '',
     // 是否自动开启 SESSION
-    'auto_start'     => true,
+    'auto_start'     => false,
 ];